changed update script to ask for wallpaper updates separately to improve speed

+++ modules/updates/update_hyprnosis.sh
@@ -26,16 +26,16 @@ header
 spin "Fetching updates for hyprnosis" git -C "$INSTALL_DIR" fetch origin
 spin "Resetting repo to main branch" git -C "$INSTALL_DIR" reset --hard origin/main
 
-spin "Updating wallpapers" git clone --depth 1 https://github.com/tyvren/hyprnosis-wallpapers.git /tmp/wallpapers
-rm -rf /tmp/wallpapers/.git
-rm -rf /tmp/wallpapers/README.md
-cp -r /tmp/wallpapers/. "$INSTALL_DIR/wallpapers/"
-rm -rf /tmp/wallpapers
-
 prompt "Updating Quickshell config"
 cp -r "$CONFIG_DIR/quickshell" "$LOCAL_CONFIG"
 
-gum confirm "Update complete. Press enter to close."
-
-pkill walker
-systemctl --user restart walker.service elephant.service
+wall_update_choice=$(gum confirm "Update hyprnosis wallpapers?")
+if $wall_update_choice; then
+  spin "Updating wallpapers" git clone --depth 1 https://github.com/tyvren/hyprnosis-wallpapers.git /tmp/wallpapers
+  rm -rf /tmp/wallpapers/.git
+  rm -rf /tmp/wallpapers/README.md
+  cp -r /tmp/wallpapers/. "$INSTALL_DIR/wallpapers/"
+  rm -rf /tmp/wallpapers
+else
+  gum confirm "Update complete. Press enter to close."
+fi
